requirements.txt
openai
python-dotenv
pyodbc
beautifulsoup4